requirements.txt
Django==2.1.3
django-bootstrap-form==3.4
djangorestframework==3.9.1
pkg-resources==0.0.0
pytz==2018.7
